requirements.txt
dnspython
whois